fix async access token
