+++ src/qianfan/resources/requestor/openapi_requestor.py
@@ -38,6 +38,7 @@ from qianfan.resources.auth.iam import iam_sign
 from qianfan.resources.auth.oauth import Auth
 from qianfan.resources.requestor.base import (
     BaseAPIRequestor,
+    _async_check_if_status_code_is_200,
     _check_if_status_code_is_200,
     _with_latency,
 )
@@ -116,7 +117,8 @@ class QfAPIRequestor(BaseAPIRequestor):
                         token_refreshed = True
                         self._auth.refresh_access_token()
                         self._add_access_token(request)
-                        responses = self._client.request_stream(request)
+                        with self._rate_limiter:
+                            responses = self._client.request_stream(request)
                         continue
                     raise
 
@@ -159,6 +161,50 @@ class QfAPIRequestor(BaseAPIRequestor):
             return await func(*args, **kwargs)
 
         return retry_wrapper
+
+    @_with_latency
+    async def _async_request_stream(
+        self,
+        request: QfRequest,
+        data_postprocess: Callable[[QfResponse], QfResponse] = lambda x: x,
+    ) -> AsyncIterator[QfResponse]:
+        """
+        async stream request
+        """
+        async with self._rate_limiter:
+            responses = self._client.arequest_stream(request)
+        token_refreshed = False
+        async for body, resp in responses:
+            _async_check_if_status_code_is_200(resp)
+            body_str = body.decode("utf-8")
+            if body_str.strip() == "":
+                continue
+            if not body_str.startswith(Consts.STREAM_RESPONSE_PREFIX):
+                try:
+                    # the response might be error message in json format
+                    json_body: Dict[str, Any] = json.loads(body_str)
+                    self._check_error(json_body)
+                except json.JSONDecodeError:
+                    # the response is not json format, ignore and raise RequestError
+                    pass
+                except errors.AccessTokenExpiredError:
+                    if not token_refreshed:
+                        token_refreshed = True
+                        await self._auth.arefresh_access_token()
+                        await self._async_add_access_token(request)
+                        async with self._rate_limiter:
+                            responses = self._client.arequest_stream(request)
+                        continue
+                    raise
+                raise errors.RequestError(
+                    f"got unexpected stream response from server: {body_str}"
+                )
+            body_str = body_str[len(Consts.STREAM_RESPONSE_PREFIX) :]
+            json_body = json.loads(body_str)
+            parsed = self._parse_async_response(json_body, resp)
+            parsed.request = QfRequest.from_aiohttp(resp.request_info)
+            parsed.request.json_body = copy.deepcopy(request.json_body)
+            yield data_postprocess(parsed)
 
     def _check_error(self, body: Dict[str, Any]) -> None:
         """
